+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Changelog
+### Unreleased
+* Fix UI freezing when loading source by dispatching the loading in the global queue instead of main queue
 
 ### Version 5.0.2
 * Fix crash when RCTVideo's superclass doesn't observe the keyPath 'frame' (iOS) [#1720](https://github.com/react-native-community/react-native-video/pull/1720)